app/requirements.txt
Flask
pdf2image
pytesseract
python-docx
